push model.GetInfoForBytes -> channels/app
